[FIX] Added checks for game ownership before operations | game card options (delete & edit)

+++ gamenest/src/main/java/com/gamenest/controller/GameController.java
@@ -2,27 +2,30 @@ package com.gamenest.controller;
 
 import java.util.List;
 import jakarta.validation.Valid;
-
+import lombok.RequiredArgsConstructor;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import com.gamenest.dto.game.GameRequest;
 import com.gamenest.dto.game.UpdateGameRequest;
+import com.gamenest.dto.user.UserRequest;
+import com.gamenest.exception.ResourceOwnershipException;
 import com.gamenest.service.interfaces.GameService;
+import com.gamenest.service.interfaces.UserService;
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/v1/games")
 @Tag(name = "Game API", description = "Operations pertaining to games in the application")
 public class GameController {
 
     private final GameService gameService;
-
-    public GameController(GameService gameService) {
-        this.gameService = gameService;
-    }
+    private final UserService userService;
 
     @Operation(summary = "Create a new game", description = "Creates a new game and associates it with the currently authenticated user.")
     @PostMapping
@@ -36,6 +39,13 @@ public class GameController {
     public ResponseEntity<GameRequest> updateGame(
             @PathVariable Long gameId,
             @RequestBody UpdateGameRequest updateGameRequest) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        UserRequest user = userService.getByUserName(username);
+        GameRequest game = gameService.getGameById(gameId);
+        if (game.getOwner().getId() != user.getId())
+            throw new ResourceOwnershipException("You are not the owner of this resource");
+
         GameRequest updatedGame = gameService.updateGame(gameId, updateGameRequest);
         return ResponseEntity.ok(updatedGame);
     }
@@ -61,6 +71,13 @@ public class GameController {
     @Operation(summary = "Delete a game", description = "Deletes a game by its ID.")
     @DeleteMapping("/{gameId}")
     public ResponseEntity<Void> deleteGame(@PathVariable Long gameId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        UserRequest user = userService.getByUserName(username);
+        GameRequest game = gameService.getGameById(gameId);
+        if (game.getOwner().getId() != user.getId())
+            throw new ResourceOwnershipException("You are not the owner of this resource");
+
         gameService.deleteGame(gameId);
         return ResponseEntity.noContent().build();
     }
